Move the auth provider
