+++ src/components/site-header/SiteHeader.tsx
@@ -2,8 +2,8 @@ import { Link } from "gatsby";
 import React from "react";
 import styled from "styled-components";
 
-import { useAuth0 } from "../../auth/react-auth0-spa";
 import Colors from "../../design-system/Colors";
+import { useAuth0 } from "../../providers/auth/react-auth0-spa";
 import Logo from "./Logo";
 
 const Nav = styled.nav`
