+++ src/pages/impact.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 
-import AuthWall from "../auth/AuthWall";
 import PageInfo from "../components/site-metadata/PageInfo";
 import ResponseImpactPage from "../page-response-impact/ResponseImpactPage";
+import AuthWall from "../providers/auth/AuthWall";
 
 // eslint-disable-next-line react/display-name
 export default () => (
