+++ src/pages/inspect-model.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 
-import AuthWall from "../auth/AuthWall";
 import PageInfo from "../components/site-metadata/PageInfo";
 import ModelInspectionPage from "../page-model-inspection/";
+import AuthWall from "../providers/auth/AuthWall";
 
 // eslint-disable-next-line react/display-name
 export default () => (
