+++ src/providers/auth/AuthWall.tsx
@@ -2,7 +2,7 @@ import { useLocation } from "@reach/router";
 import { navigate } from "gatsby";
 import React, { useEffect } from "react";
 
-import Loading from "../design-system/Loading";
+import Loading from "../../design-system/Loading";
 import { useAuth0 } from "./react-auth0-spa";
 
 const AuthWall: React.FC = (props) => {
