+++ src/providers/site-provider/SiteProvider.tsx
@@ -2,8 +2,6 @@ import { navigate } from "gatsby";
 import React, { useReducer, useState } from "react";
 import { ToastProvider } from "react-toast-notifications";
 
-import AppAuth0ClientPromise from "../../auth/AppAuth0ClientPromise";
-import { Auth0Provider } from "../../auth/react-auth0-spa";
 import {
   FacilityContext,
   rtDataReducer,
@@ -12,6 +10,8 @@ import { LocaleDataProvider } from "../../contexts/locale-data-context";
 import { ScenarioProvider } from "../../contexts/scenario-context";
 import Toast from "../../design-system/Toast";
 import { Facility } from "../../page-multi-facility/types";
+import AppAuth0ClientPromise from "../auth/AppAuth0ClientPromise";
+import { Auth0Provider } from "../auth/react-auth0-spa";
 import { FeatureFlagsProvider } from "../feature-flags";
 
 // A function that routes the user to the right place after login
